Refactored casem page translations into CaseMTranslator class
